restructured repo with gulp
