+++ demo/index.html
@@ -5,7 +5,7 @@
         <link rel="stylesheet" href="http://cdn.leafletjs.com/leaflet/v1.0.0-beta.2/leaflet.css" />
         <link rel="stylesheet" href="demo.css" />
         <script src="http://cdn.leafletjs.com/leaflet/v1.0.0-beta.2/leaflet.js"></script>
-        <script src="../leaflet.pm.js"></script>
+        <script src="../dist/leaflet.pm.js"></script>
     </head>
     <body>
