+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "leaflet.pm",
+  "version": "1.0.0",
+  "description": "A Leaflet Plugin For Editing Geometry Layers in Leaflet 1.0",
+  "main": "leaflet.pm.js",
+  "dependencies": {
+    "gulp": "^3.9.0",
+    "gulp-concat": "^2.6.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-uglify": "^1.5.1"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/codeofsumit/leaflet.pm.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/codeofsumit/leaflet.pm/issues"
+  },
+  "homepage": "https://github.com/codeofsumit/leaflet.pm#readme"
+}
