+++ src/L.PM.js
@@ -1,0 +1,8 @@
+/**
+*
+* A Leaflet Plugin For Editing Geometry Layers in Leaflet 1.0
+* by Sumit Kumar (@TweetsOfSumit)
+* Github Repo: https://github.com/codeofsumit/leaflet.pm
+*/
+
+L.PM = L.PM || {};
